refactor: extract ComputeSettings panel from SettingsDialog (#672)
Second per-tab panel split. The compute tab (#374) — the per-machine Python
interpreter override + live probe — is self-contained (its state and handlers
only touch api.compute.*), so it lifts whole into ComputeSettings.svelte that
owns its state, loads on mount, and is mounted by the dialog for the "compute"
tab. SettingsDialog 2052 → 1912.

Compute-specific CSS (.path-row / .probe-* / .action-row / the compute-only
.action-btn.primary) moved into the child and deleted from the dialog; the
shared .action-btn / .link-btn / .field / .hint rules stay (other panels use
them) and the child carries its own scoped copies per the app's per-dialog
convention.

Behavior note: like the sites panel, settings now load lazily when the compute
tab is opened (the child's onMount) rather than eagerly on dialog open.

Tests (5): components/ComputeSettings.test.ts mocks api.compute and pins the
probe-on-mount + version line, the probe error state, Browse → set-path + re-
probe, Save gated on dirty state → setPythonSettings, and Clear override.

Verified: pnpm lint (0 errors), full suite (2990 passed, +5), electron-forge
build + boot smoke.
