+++ src/renderer/lib/components/SettingsDialog.svelte
@@ -52,6 +52,7 @@
   } from '../../../shared/skills/menu-config';
   import { SKILL_MENUS } from '../../../shared/skills/types';
   import SitesSettings from './SitesSettings.svelte';
+  import ComputeSettings from './ComputeSettings.svelte';
 
   interface Props {
     onApplyEditor: (s: EditorSettings) => void;
@@ -456,66 +457,8 @@
   let toolModelOverrides = $state<Record<string, string>>({});
   const allTools: ThinkingToolInfo[] = getAllToolInfos();
 
-  // Compute (#374): per-machine Python interpreter override.
-  let pythonPathInput = $state('');
-  /** What's saved to disk; used to detect dirty state. */
-  let pythonPathSaved = $state('');
-  let pythonProbe = $state<{ ok: boolean; path: string; version?: string; error?: string } | null>(null);
-  let pythonProbing = $state(false);
-
-  async function loadComputeSettings(): Promise<void> {
-    try {
-      const s = await api.compute.getPythonSettings();
-      pythonPathInput = s.pythonPath;
-      pythonPathSaved = s.pythonPath;
-      // Probe whatever the resolver would currently pick so the
-      // status line reflects the live state, not just the override.
-      await refreshPythonProbe();
-    } catch (e) {
-      console.error('[settings] failed to load python settings:', e);
-    }
-  }
-
-  async function refreshPythonProbe(): Promise<void> {
-    pythonProbing = true;
-    try {
-      // Empty `pythonPathInput` → probe the resolver's active pick
-      // (env var or `python3`). Non-empty → probe the input directly
-      // so the status line shows whether the candidate would work.
-      pythonProbe = await api.compute.probePython(pythonPathInput.trim() || undefined);
-    } catch (e) {
-      pythonProbe = { ok: false, path: pythonPathInput, error: e instanceof Error ? e.message : String(e) };
-    } finally {
-      pythonProbing = false;
-    }
-  }
-
-  async function browsePythonInterpreter(): Promise<void> {
-    const picked = await api.compute.browsePython();
-    if (!picked) return;
-    pythonPathInput = picked;
-    // Probe immediately so the user gets instant feedback on whether
-    // the picked file is actually a runnable Python.
-    await refreshPythonProbe();
-  }
-
-  async function savePythonPath(): Promise<void> {
-    try {
-      await api.compute.setPythonSettings({ pythonPath: pythonPathInput.trim() });
-      pythonPathSaved = pythonPathInput.trim();
-      await refreshPythonProbe();
-    } catch (e) {
-      pythonProbe = { ok: false, path: pythonPathInput, error: e instanceof Error ? e.message : String(e) };
-    }
-  }
-
-  async function restartPythonKernelFromSettings(): Promise<void> {
-    try {
-      await api.compute.restartPythonKernel();
-    } catch (e) {
-      console.error('[settings] failed to restart python kernel:', e);
-    }
-  }
+  // Compute (#374): the Python-interpreter panel now lives in
+  // ComputeSettings.svelte (self-contained).
 
   onMount(async () => {
     try {
@@ -533,7 +476,6 @@
     }
     await loadBibliographySettings();
     await loadExcerptSettings();
-    await loadComputeSettings();
     await loadSkills();
     try {
       const ingest = await api.sources.getIngestSettings();
@@ -1203,89 +1145,7 @@
           </div>
 
         {:else if activeTab === 'compute'}
-          <div class="field">
-            <label for="python-path">Python interpreter</label>
-            <p class="hint">
-              Path to the Python executable Minerva should use for cell
-              execution. Leave empty to fall back to the
-              <code>MINERVA_PYTHON</code> environment variable, then to
-              <code>python3</code> on <code>$PATH</code>. Stored
-              per-machine — different projects on this machine share the
-              same interpreter.
-            </p>
-            <div class="path-row">
-              <input
-                id="python-path"
-                type="text"
-                bind:value={pythonPathInput}
-                placeholder="/Users/you/.minerva-venv/bin/python"
-                spellcheck="false"
-                autocomplete="off"
-                autocapitalize="off"
-              />
-              <button
-                class="action-btn"
-                onclick={() => { void browsePythonInterpreter(); }}
-                disabled={pythonProbing}
-              >
-                Browse…
-              </button>
-              <button
-                class="action-btn"
-                onclick={() => { void refreshPythonProbe(); }}
-                disabled={pythonProbing}
-                title="Test the interpreter — runs `python --version`"
-              >
-                {pythonProbing ? 'Probing…' : 'Probe'}
-              </button>
-            </div>
-
-            {#if pythonProbe}
-              <div class="probe-result" class:probe-ok={pythonProbe.ok} class:probe-error={!pythonProbe.ok}>
-                {#if pythonProbe.ok}
-                  <strong>{pythonProbe.version}</strong>
-                  <span class="probe-path">at <code>{pythonProbe.path}</code></span>
-                {:else}
-                  <strong>Couldn't run interpreter:</strong>
-                  <span class="probe-path">{pythonProbe.error}</span>
-                {/if}
-              </div>
-            {/if}
-
-            <div class="action-row">
-              <button
-                class="action-btn primary"
-                onclick={() => { void savePythonPath(); }}
-                disabled={pythonProbing || pythonPathInput.trim() === pythonPathSaved}
-              >
-                Save
-              </button>
-              <button
-                class="action-btn"
-                onclick={() => { void restartPythonKernelFromSettings(); }}
-                title="Apply the new interpreter to a fresh kernel — wipes namespace state"
-              >
-                Save &amp; Restart Kernel
-              </button>
-              {#if pythonPathSaved}
-                <button
-                  class="link-btn"
-                  onclick={() => { pythonPathInput = ''; void savePythonPath(); }}
-                >
-                  Clear override
-                </button>
-              {/if}
-            </div>
-
-            <p class="hint">
-              Tip: a venv at <code>~/.minerva-venv/bin/python</code> with
-              <code>pandas</code>, <code>matplotlib</code>, and
-              <code>pillow</code> installed gives you the full rich-output
-              pipeline. After changing the interpreter, click
-              <em>Save &amp; Restart Kernel</em> so the next cell runs
-              against the new env.
-            </p>
-          </div>
+          <ComputeSettings />
 
         {:else if activeTab === 'ai'}
           <div class="field">
@@ -1934,56 +1794,6 @@
   }
   .skill-err-label {
     font-weight: 600;
-  }
-
-  /* Compute panel — Python interpreter row + probe status (#374). */
-  .path-row {
-    display: flex;
-    gap: 6px;
-    align-items: center;
-    margin: 6px 0;
-  }
-  .path-row input[type="text"] {
-    flex: 1;
-    min-width: 0;
-    font-family: var(--font-mono, ui-monospace, monospace);
-    font-size: 12px;
-  }
-  .probe-result {
-    padding: 6px 10px;
-    border-left: 3px solid var(--border);
-    background: var(--bg-button);
-    font-size: 12px;
-    margin: 6px 0;
-    border-radius: 0 3px 3px 0;
-  }
-  .probe-result.probe-ok { border-left-color: var(--accent); }
-  .probe-result.probe-error { border-left-color: var(--accent); }
-  .probe-result strong { display: block; margin-bottom: 2px; }
-  .probe-result .probe-path {
-    color: var(--text-muted);
-    font-size: 11px;
-  }
-  .probe-result code {
-    font-size: 11px;
-    background: var(--bg);
-    padding: 1px 4px;
-    border-radius: 2px;
-  }
-  .action-row {
-    display: flex;
-    gap: 6px;
-    align-items: center;
-    margin: 8px 0;
-  }
-  .action-btn.primary {
-    background: var(--accent);
-    color: var(--bg);
-    border-color: var(--accent);
-    font-weight: 500;
-  }
-  .action-btn.primary:hover:not(:disabled) {
-    filter: brightness(1.1);
   }
 
   .section-intro code {
